+++ CHANGELOG.md
@@ -1,5 +1,7 @@
+### 1.7.7: September 7th, 2021
 * Fix cancel button in admin did not open the confirm prompt
 * Alternative 1 calculation now default
+* Free values can't be above capacity nor below 0
 
 ### 1.7.6: July 15th, 2021
 * Alternative calculation of free place